+++ requirements.txt
@@ -1,3 +1,5 @@
 flask
 numpy
 pandas
+serverless-wsgi
+scikit-learn
